# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/properties": {
      "get": {
        "operationId": "PropertiesController_findAll",
        "summary": "Mirrors PropertyRepository.all()/.discoverable() — pass discoverableOnly=true for the latter.",
        "parameters": [
          {
            "name": "discoverableOnly",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ]
      },
      "post": {
        "operationId": "PropertiesController_create",
        "summary": "Publish a listing under the caller’s own broker profile (B03 save).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDto"
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/properties/photos/{photoId}": {
      "get": {
        "operationId": "PropertiesController_findPhoto",
        "summary": "Bytes behind an `api:<id>` photoAssets entry. Public, like the listing itself.",
        "parameters": [
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "properties"
        ]
      }
    },
    "/properties/{id}": {
      "get": {
        "operationId": "PropertiesController_findById",
        "summary": "Mirrors PropertyRepository.byId.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDto"
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ]
      },
      "patch": {
        "operationId": "PropertiesController_update",
        "summary": "Edit one of the caller’s own listings. 403 for someone else’s.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDto"
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PropertiesController_close",
        "summary": "Withdraw a listing. Soft delete to CLOSED so client contact history keeps resolving it.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDto"
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/reviews": {
      "get": {
        "operationId": "ReviewsController_all",
        "parameters": [
          {
            "name": "onlyPublic",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ReviewsController_create",
        "summary": "Port of ReviewEligibilityService — 403s with a `reason` (noContact/notReached/alreadyReviewed/notOwner) when not eligible.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reviews/broker/{brokerId}": {
      "get": {
        "operationId": "ReviewsController_byBroker",
        "parameters": [
          {
            "name": "brokerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "onlyPublic",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Woutalma Keur API",
    "description": "Phase 1: broker/property read, ranked search, contact logging, review eligibility.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GeoPointDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
...
      "PropertyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "brokerId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "HOUSE",
              "LAND",
              "STUDIO",
              "ROOM"
            ]
          },
          "transaction": {
            "type": "string",
            "enum": [
              "RENT",
              "SALE"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "CFA francs, integer."
          },
          "surface": {
            "type": "number",
            "nullable": true
          },
          "rooms": {
            "type": "number",
            "nullable": true
          },
          "position": {
            "$ref": "#/components/schemas/GeoPointDto"
          },
          "neighbourhood": {
            "type": "string"
          },
          "photoAssets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "CLOSED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "isDiscoverable": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "brokerId",
          "kind",
          "transaction",
          "title",
          "description",
          "price",
          "position",
          "neighbourhood",
          "photoAssets",
          "status",
          "createdAt",
          "isDiscoverable"
        ]
      },
      "UploadPhotoDto": {
        "type": "object",
        "properties": {
          "mimeType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          },
          "dataBase64": {
            "type": "string",
            "description": "Base64-encoded image bytes. Decoded size must not exceed 163840 bytes — compress before sending."
          }
        },
        "required": [
          "mimeType",
          "dataBase64"
        ]
      },
      "CreatePropertyDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "HOUSE",
              "LAND",
              "STUDIO",
              "ROOM"
            ]
          },
          "transaction": {
            "type": "string",
            "enum": [
              "RENT",
              "SALE"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "CFA francs, integer."
          },
          "surface": {
            "type": "number",
            "nullable": true
          },
          "rooms": {
            "type": "number",
            "nullable": true
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "neighbourhood": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "CLOSED"
            ]
          },
          "photoAssets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "newPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadPhotoDto"
            }
          }
        },
        "required": [
          "kind",
          "transaction",
          "title",
          "price",
          "latitude",
          "longitude",
          "neighbourhood"
        ]
      },
      "UpdatePropertyDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "HOUSE",
              "LAND",
              "STUDIO",
              "ROOM"
            ]
          },
          "transaction": {
            "type": "string",
            "enum": [
              "RENT",
              "SALE"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "CFA francs, integer."
          },
          "surface": {
            "type": "number",
            "nullable": true
          },
          "rooms": {
            "type": "number",
            "nullable": true
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "neighbourhood": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "CLOSED"
            ]
          },
          "photoAssets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "newPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadPhotoDto"
            }
          }
        }
      },
...
      "ReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "brokerId": {
            "type": "string"
          },
          "contactId": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "responsiveness": {
            "type": "number",
            "nullable": true
          },
          "accuracy": {
            "type": "number",
            "nullable": true
          },
          "courtesy": {
            "type": "number",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "moderation": {
            "type": "string",
            "enum": [
              "PENDING",
              "PUBLISHED",
              "REJECTED"
            ]
          },
          "brokerReply": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "isPublic": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "brokerId",
          "contactId",
          "rating",
          "moderation",
          "createdAt",
          "isPublic"
        ]
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "contactId": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "responsiveness": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "accuracy": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "courtesy": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "contactId",
          "rating"
        ]
      },
...
    }
  }
}